Servant.CLI

Documentation

class RunClient m => HasClient (m :: Type -> Type) api where

This class lets us define how each API combinator influences the creation of an HTTP request.

Unless you are writing a new backend for servant-client-core or new combinators that you want to support client-generation, you can ignore this class.

Associated Types

type Client (m :: Type -> Type) api :: Type

Methods

clientWithRoute :: Proxy m -> Proxy api -> Request -> Client m api

hoistClientMonad :: Proxy m -> Proxy api -> (forall x. mon x -> mon' x) -> Client mon api -> Client mon' api

clientPStruct :: HasCLI m api => Proxy api -> Proxy m -> PStruct (CLI m api)

parseClient :: HasCLI ClientM api => Proxy api -> IO (CLI ClientM api)

class ParseBody a where

Minimal complete definition

Nothing

Methods

parseBody :: Parser a

parseBody :: (Typeable a, Read a) => Parser a

defaultParseBody :: forall a. Typeable a => ReadM a -> Parser a

Re-export

class ToSample a where

The class that lets us display a sample input or output in the supported content-types when generating documentation for endpoints that either:

expect a request body, or

return a non empty response body

Example of an instance:

{-# LANGUAGE DeriveGeneric #-}
{-# LANGUAGE OverloadedStrings #-}

import Data.Aeson
import Data.Text
import GHC.Generics

data Greet = Greet { _msg :: Text }
  deriving (Generic, Show)

instance FromJSON Greet
instance ToJSON Greet

instance ToSample Greet where
  toSamples _ = singleSample g

    where g = Greet "Hello, haskeller!"

You can also instantiate this class using toSamples instead of toSample: it lets you specify different responses along with some context (as ErrorMessage) that explains when you're supposed to get the corresponding response.

Minimal complete definition

Nothing

Methods

toSamples :: Proxy a -> [(Text, a)]

class ToCapture (c :: k) where

The class that helps us automatically get documentation for URL captures.

Example of an instance:

instance ToCapture (Capture "name" Text) where
  toCapture _ = DocCapture "name" "name of the person to greet"

Methods

toCapture :: Proxy c -> DocCapture

data DocCapture

A type to represent captures. Holds the name of the capture and a description.

Write a ToCapture instance for your captured types.

Constructors

DocCapture

Fields

_capSymbol :: String

_capDesc :: String

class ToParam (t :: k) where

The class that helps us automatically get documentation for GET (or other Method) parameters.

Example of an instance:

instance ToParam (QueryParam' mods "capital" Bool) where
  toParam _ =
    DocQueryParam "capital"
                  ["true", "false"]
                  "Get the greeting message in uppercase (true) or not (false). Default is false."

Methods

toParam :: Proxy t -> DocQueryParam

data DocQueryParam

A type to represent a GET (or other possible Method) parameter from the Query String. Holds its name, the possible values (leave empty if there isn't a finite number of them), and a description of how it influences the output or behavior.

Write a ToParam instance for your GET parameter types

Constructors

DocQueryParam

Fields

_paramName :: String

_paramValues :: [String]

_paramDesc :: String

_paramKind :: ParamKind