Servant.CLI.HasCLI

Description

Main module providing underlying functionality for the command line interface parser for servant API clients.

For the most part, you can ignore this module unless you're adding new API combinators.

Documentation

class HasCLI m api where

Typeclass defining how each API combinator influences how a server can be interacted with using command line options.

Note that query parameters and captures all require servant-docs annotation instances, to allow for proper help messages.

Unless you are adding new combinators to be used with APIs, you can ignore this class.

Associated Types

type CLIParam (m :: Type -> Type) (api :: Type) :: [Type]

List of extra parameters that must be generated at runtime to produce the desired request.

For most routes this will be an empty list, but it is non-empty for StreamBody, BasicAuth, and AuthProtect.

type CLIResult (m :: Type -> Type) (api :: Type) :: Type

The parsed type of the client request response. Usually this will be a bunch of nested Eithers for every API endpoint, nested according to the :<|>s in the API.

type CLIHandler (m :: Type -> Type) (api :: Type) (r :: Type) :: Type

The type of a data structure to conveniently handle the results of all pontential endpoints. This is useful because it is often tedious to handle the bunch of nested Eithers that CLIResult has.

It essentially lets you specify how to combine each potential endpoint's response into a single output value.

Usually this will be a bunch of nested :<|>s which handle each endpoint, according to the :<|>s in the API. It mirrors the structure of Client and ServerT.

Used with functions like parseHandleClient.

Methods

cliPStruct_ :: Proxy m -> Proxy api -> Rec m (CLIParam m api) -> PStruct (Request -> m (CLIResult m api))

Generate a PStruct showing how to modify a Request and perform an action, given an API and underlying monad. Only meant for internal use; should be used through cliPStruct instead.

Takes a Rec of actions to generate required items. Pass in RNil if no parameters are expected (that is, if CLIParam m api is an empty list). The actions will only be run if they are needed.

cliHandler :: Proxy m -> Proxy api -> CLIHandler m api r -> CLIResult m api -> r

Handle all the possibilities in a CLIResult, by giving the appropriate CLIHandler.

Utility

splitRec :: forall as bs g. RecApplicative as => Rec g (as ++ bs) -> (Rec g as, Rec g bs)

Un-append two Recs.