Servant.CLI.Internal

Documentation

class HasCLI m api where

Typeclass defining how each API combinator influences how a server can be interacted with using command line options.

Unless you are adding new combinators to be used with APIs, you can ignore this class.

Associated Types

type CLI (m :: Type -> Type) (api :: Type) :: Type

Methods

clientPStruct_ :: Proxy m -> Proxy api -> PStruct (Request -> CLI m api)

class ParseBody a where

A helper class for defining directly how to parse request bodies. This allows more complex parsing of bodies.

Minimal complete definition

Nothing

Methods

parseBody :: Parser a

parseBody :: (Typeable a, Read a) => Parser a

defaultParseBody

Arguments

:: String
type specification

-> ReadM a
parser

-> Parser a