Servant.CLI.ParseBody

Description

Provides the interface for ParseBody, a helper class for defining directly how to parse request bodies.

Documentation

class ParseBody a where

A helper class for defining directly how to parse request bodies. This allows more complex parsing of bodies.

You need an instance of this for every type you use with ReqBody.

Minimal complete definition

Nothing

Methods

parseBody :: Parser a

parseBody :: (Typeable a, Read a) => Parser a

defaultParseBody

Arguments

:: String
type specification

-> ReadM a
parser

-> Parser a

Default implementation that expects a --data option.