Servant.CLI.Structure

Documentation

data OptRead :: Type -> Type where

How to "read" an option.

Constructors

ORRequired :: ReadM a -> OptRead a

OROptional :: ReadM a -> OptRead (Maybe a)

ORSwitch :: OptRead Bool

data Opt a

Query parameters are interpreted as options

Constructors

Opt

Fields

optName :: String

optDesc :: String

optMeta :: String

optVals :: Maybe (NonEmpty String)

optRead :: Coyoneda OptRead a

data Arg a

Captures are interpreted as arguments

Constructors

Arg

Fields

argName :: String

argDesc :: String

argMeta :: String

argRead :: ReadM a

data MultiArg :: Type -> Type where

Constructors

MultiArg :: Arg a -> MultiArg [a]

type Captures = Day Arg PStruct :+: Day MultiArg EndpointMap

data Endpoint a

Endpoint arguments and body.

TODO: add things like status etc.

Constructors

Endpoint

Fields

epStruct :: Day (Ap Opt) Parser a

data EndpointMap a

Constructors

EPM

Fields

epmGiven :: Map Method (Endpoint a)

epmRaw :: Maybe (Day ((:~:) Method) Endpoint a)

data PStruct a

Structure for a parser of a given value that may use items from captures and arguments.

Constructors

PStruct

Fields

psInfo :: [String]

psComponents :: Map String (PStruct a)

path components

psCaptures :: Maybe (Captures a)

captures

psEndpoints :: EndpointMap a

data PStructF (a :: Type) r

Constructors

PStructF

Fields

psInfoF :: [[Char]]

psComponentsF :: (Map [Char] r)

psCapturesF :: (Maybe ((:+:) (Day Arg PStruct) (Day MultiArg EndpointMap) a))

psEndpointsF :: (EndpointMap a)

structParser

Arguments

:: PStruct a
The PStruct to convert.

-> InfoMod a
Modify how the top-level prompt is displayed.

-> ParserInfo a

Convert a PStruct into a command line argument parser, from the optparse-applicative library. It can be run with execParser.

It takes options on how the top-level prompt is displayed when given "--help"; it can be useful for adding a header or program description. Otherwise, just use mempty.

structParser_

Arguments

:: PStruct a

-> Bool
add helper

-> [String]
root path

-> InfoMod a
modify top level

-> ParserInfo a

Low-level implementation of structParser.

Combining

altPStruct :: PStruct a -> PStruct a -> PStruct a

Combine two PStructs, preferring the left hand side for conflicts. If the left hand has a capture, the right hand's components are ignored. If the left hand has a raw endpoint, the right hand's endpoints are ignored.

altEPM :: EndpointMap a -> EndpointMap a -> EndpointMap a

Combine two EndpointMaps, preferring the left hand side for conflicts. If the left hand has a raw endpoint, the right hand's endpoints are ignored.

branch :: PStruct a -> PStruct b -> PStruct (Either a b) infixr 3

Creating

($:>) :: String -> PStruct a -> PStruct a infixr 4

Shift by a path component.

(%:>) :: Parser a -> PStruct (a -> b) -> PStruct b infixr 4

Add a request body to all endpoints. NOTE!!!!! UNDEFINED BEHAVIOR IF DONE MORE THAN ONCE?

(?:>) :: Opt a -> PStruct (a -> b) -> PStruct b infixr 4

Add a command-line option to all endpoints.

(#:>) :: Arg a -> PStruct (a -> b) -> PStruct b infixr 4

Add a single argument praser.

(##:>) :: Arg a -> PStruct ([a] -> b) -> PStruct b infixr 4

Add a repeating argument parser.

note :: String -> PStruct a -> PStruct a infixr 4

Add a note.

endpoint :: Method -> a -> PStruct a

Create an endpoint action.

rawEndpoint :: (Method -> a) -> PStruct a

Create a raw endpoint.

Readers

orRequired :: ReadM a -> Coyoneda OptRead a

Helper to lift a ReadM into something that can be used with optRead.

orOptional :: ReadM a -> Coyoneda OptRead (Maybe a)

Helper to lift an optional ReadM into something that can be used with optRead.

orSwitch :: Coyoneda OptRead Bool

An optRead that is on-or-off.