Servant.CLI

Description

Parse command line arguments into a servant client, from a servant API.

Mainly used through parseClient and parseHandleClient. parseClient returns a servant client action that returns nested Eithers for every endpoint, but parseHandleClient allows you to conveniently specify how you want to combine each endpoint entry into a single result.

Documentation

parseClient

Arguments

:: HasCLI m api

=> Proxy api
API

-> Proxy m
Client monad

-> Rec m (CLIParam m api)
Extra parameters

-> InfoMod (m (CLIResult m api))
Options for top-level display

-> IO (m (CLIResult m api))

Parse a servant client; the result can be run. A good choice of m is ClientM.

Returns the request response, which is usually a layer of Either for every endpoint branch. You can find the response type directly by using typed holes or asking ghci with :t or :kind! forall m. CLIResult m MyAPI. Because it might be tedious handling nested Eithers, see parseHandleClient for a way to handle each potential branch in a convenient way.

Takes a Rec of actions to generate required items. Pass in RNil if no parameters are expected (that is, if CLIParam m api is an empty list), or use parseClient'. The actions will only be run if they are needed.

Takes options on how the top-level prompt is displayed when given "--help"; it can be useful for adding a header or program description. Otherwise, just use mempty.

parseHandleClient

Arguments

:: (HasCLI m api, Functor m)

=> Proxy api
API

-> Proxy m
Client monad

-> Rec m (CLIParam m api)
Extra parameters

-> InfoMod (m (CLIResult m api))
Options for top-level display

-> CLIHandler m api r
Handler

-> IO (m r)

Parse a server client, like parseClient. However, instead of that client action returning the request response, instead use a CLIHandler to handle every potential request response. It essentially lets you specify how to combine each potential endpoint's response into a single output value.

The handler is usually a :<|> for every endpoint branch. You can find it by using typed holes or asking ghci with :t or :kind! forall m r. CLIHandler m MyAPI r.

Takes a Rec of actions to generate required items. Pass in RNil if no parameters are expected (that is, if CLIParam m api is an empty list), or use parseHandleClient'. The actions will only be run if they are needed.

Takes options on how the top-level prompt is displayed when given "--help"; it can be useful for adding a header or program description. Otherwise, just use mempty.

parseClient'

Arguments

:: (HasCLI m api, CLIParam m api ~ '[])

=> Proxy api
API

-> Proxy m
Client monad

-> InfoMod (m (CLIResult m api))
Options for top-level display

-> IO (m (CLIResult m api))

A version of parseClient that works when the client action requires no extra parameters.

parseHandleClient'

Arguments

:: (HasCLI m api, Functor m, CLIParam m api ~ '[])

=> Proxy api
API

-> Proxy m
Client monad

-> InfoMod (m (CLIResult m api))
Options for top-level display

-> CLIHandler m api r
Handler

-> IO (m r)

A version of parseHandleClient that works when the client action requires

Typeclasses

class HasCLI m api where

Typeclass defining how each API combinator influences how a server can be interacted with using command line options.

Note that query parameters and captures all require servant-docs annotation instances, to allow for proper help messages.

Unless you are adding new combinators to be used with APIs, you can ignore this class.

Minimal complete definition

cliPStruct_, cliHandler

Associated Types

type CLIParam (m :: Type -> Type) (api :: Type) :: [Type]

List of extra parameters that must be generated at runtime to produce the desired request.

For most routes this will be an empty list, but it is non-empty for StreamBody, BasicAuth, and AuthProtect.

type CLIResult (m :: Type -> Type) (api :: Type) :: Type

The parsed type of the client request response. Usually this will be a bunch of nested Eithers for every API endpoint, nested according to the :<|>s in the API.

type CLIHandler (m :: Type -> Type) (api :: Type) (r :: Type) :: Type

The type of a data structure to conveniently handle the results of all pontential endpoints. This is useful because it is often tedious to handle the bunch of nested Eithers that CLIResult has.

It essentially lets you specify how to combine each potential endpoint's response into a single output value.

Usually this will be a bunch of nested :<|>s which handle each endpoint, according to the :<|>s in the API. It mirrors the structure of Client and ServerT.

Used with functions like parseHandleClient.

Methods

cliHandler :: Proxy m -> Proxy api -> CLIHandler m api r -> CLIResult m api -> r

Handle all the possibilities in a CLIResult, by giving the appropriate CLIHandler.

Lower-level

cliPStruct

Arguments

:: HasCLI m api

=> Proxy m
Client monad

-> Proxy api
API

-> Rec m (CLIParam m api)
Extra parameters

-> PStruct (m (CLIResult m api))

Create a structure for a command line parser.

Takes a Rec of actions to generate required items. Pass in RNil if no parameters are expected (that is, if CLIParam m api is an empty list). The actions will only be run if they are needed.

cliPStruct'

Arguments

:: (HasCLI m api, CLIParam m api ~ '[])

=> Proxy m
Client monad

-> Proxy api
API

-> PStruct (m (CLIResult m api))

A version of cliPStruct that works when the client action requires no extra parameters.

structParser

Arguments

:: PStruct a
The PStruct to convert.

-> InfoMod a
Modify how the top-level prompt is displayed.

-> ParserInfo a

Convert a PStruct into a command line argument parser, from the optparse-applicative library. It can be run with execParser.

It takes options on how the top-level prompt is displayed when given "--help"; it can be useful for adding a header or program description. Otherwise, just use mempty.

Re-export

class ParseBody a where

A helper class for defining directly how to parse request bodies. This allows more complex parsing of bodies.

You need an instance of this for every type you use with ReqBody.

Minimal complete definition

Nothing

Methods

parseBody :: Parser a

parseBody :: (Typeable a, Read a) => Parser a

defaultParseBody

Arguments

:: String
type specification

-> ReadM a
parser

-> Parser a

Default implementation that expects a --data option.

class ToCapture (c :: k) where

The class that helps us automatically get documentation for URL captures.

Example of an instance:

instance ToCapture (Capture "name" Text) where
  toCapture _ = DocCapture "name" "name of the person to greet"

Methods

toCapture :: Proxy c -> DocCapture

data DocCapture

A type to represent captures. Holds the name of the capture and a description.

Write a ToCapture instance for your captured types.

Constructors

DocCapture

Fields

_capSymbol :: String

_capDesc :: String

class ToParam (t :: k) where

The class that helps us automatically get documentation for GET (or other Method) parameters.

Example of an instance:

instance ToParam (QueryParam' mods "capital" Bool) where
  toParam _ =
    DocQueryParam "capital"
                  ["true", "false"]
                  "Get the greeting message in uppercase (true) or not (false). Default is false."

Methods

toParam :: Proxy t -> DocQueryParam

data DocQueryParam

A type to represent a GET (or other possible Method) parameter from the Query String. Holds its name, the possible values (leave empty if there isn't a finite number of them), and a description of how it influences the output or behavior.

Write a ToParam instance for your GET parameter types

Constructors

DocQueryParam

Fields

_paramName :: String

_paramValues :: [String]

_paramDesc :: String

_paramKind :: ParamKind

data ParamKind

Type of GET (or other Method) parameter:

Normal corresponds to QueryParam, i.e your usual GET parameter

List corresponds to QueryParams, i.e GET parameters with multiple values

Flag corresponds to QueryFlag, i.e a value-less GET parameter

Constructors

Normal

List

Flag

data Rec (a :: u -> Type) (b :: [u]) :: forall u. (u -> Type) -> [u] -> Type where

A record is parameterized by a universe u, an interpretation f and a list of rows rs. The labels or indices of the record are given by inhabitants of the kind u; the type of values at any label r :: u is given by its interpretation f r :: *.

Constructors

RNil :: forall u (a :: u -> Type) (b :: [u]). Rec a ([] :: [u])

(:&) :: forall u (a :: u -> Type) (b :: [u]) (r :: u) (rs :: [u]). !(a r) -> !(Rec a rs) -> Rec a (r ': rs) infixr 7