Servant.CLI.HasCLI

Description

Main module providing underlying functionality for the command line interface parser for servant API clients.

For the most part, you can ignore this module unless you're adding new API combinators.

Class

class HasCLI m api ctx where

Typeclass defining how each API combinator influences how a server can be interacted with using command line options.

Note that query parameters and captures all require servant-docs annotation instances, to allow for proper help messages.

Unless you are adding new combinators to be used with APIs, you can ignore this class.

Associated Types

type CLIResult (m :: Type -> Type) (api :: Type) :: Type

The parsed type of the client request response. Usually this will be a bunch of nested Eithers for every API endpoint, nested according to the :<|>s in the API.

type CLIHandler (m :: Type -> Type) (api :: Type) (r :: Type) :: Type

The type of a data structure to conveniently handle the results of all pontential endpoints. This is useful because it is often tedious to handle the bunch of nested Eithers that CLIResult has.

It essentially lets you specify how to sort each potential endpoint's response into a single output value.

Usually this will be a bunch of nested :<|>s which handle each endpoint, according to the :<|>s in the API. It mirrors the structure of Client and ServerT.

Used with functions like parseHandleClient.

Methods

cliPStructWithContext_ :: Proxy m -> Proxy api -> Rec (ContextFor m) ctx -> PStruct (Request -> m (CLIResult m api))

Create a structure for a command line parser, which parses how to modify a Request and perform an action, given an API and underlying monad. Only meant for internal use; should be used through cliPStructWithContext instead.

Takes a Rec of actions to generate required items that cannot be passed via the command line (like authentication). Pass in RNil if no parameters are expected. The actions will only be run if they are needed.

cliHandler :: Proxy m -> Proxy api -> Proxy ctx -> CLIHandler m api r -> CLIResult m api -> r

Handle all the possibilities in a CLIResult, by giving the appropriate CLIHandler.

Context

data family ContextFor (m :: Type -> Type) :: Type -> Type

Data family associating API combinators with contexts required to run them. These typically will be actions in m that fetch/generate the required data, and will only be "run" if the user selects an endpoint that requires it through the command line interface.

newtype NamedContext m (name :: Symbol) (subContext :: [Type])

Contains a subcontext that can be descended down into using NamedContext. Mirrors NamedContext.

Useful for when you have multiple items with the same name within a context; this essentially creates a namespace for context items.

Constructors

NamedContext (Rec (ContextFor m) subContext)

descendIntoNamedContext :: forall (name :: Symbol) context subContext m. NamedContext m name subContext ∈ context => Proxy name -> Rec (ContextFor m) context -> Rec (ContextFor m) subContext

Allows you to access NamedContexts inside a context.